requirements.txt
cmap==0.7.2
entsoe_py==0.7.10
hyperopt==0.2.7
matplotlib==3.7.5
numpy==2.4.4
pandas==1.5.3
python-dotenv==1.2.2
remodels==1.0.1
scikit_learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==72.1.0
statsmodels==0.14.6
tensorflow==2.16.2
